Opaleye.TF

Documentation

Welcome to opaleye-tf, a library to query and interact with PostgreSQL databases. As the name suggests, this library builds on top of the terrific opaleye library, but provides a different API that the author believes provides more succinct code with better type inference.

The basic idea behind opaleye-tf is to "pivot" around the ideas in opaleye. The current idiomatic usage of Opaleye is to define your records as data types where each field is parameterized. Opaleye then varies all of these parameters together. opaleye-tf makes the observation that if all of these vary uniformly, then there should only be one parameter, and thus we have records that are parameterized by functors.

To take an example, let's consider a simple schema for Hackage - the repository of Haskell libraries.

data Package f =
  Package { packageName :: Col f PGText
          , packageAuthor :: Col f 'PGInteger
          , packageMaintainerId :: Col f ('PGNull 'PGInteger)
          }

data User f =
  User { userId :: Col f 'PGInteger
       , userName :: Col f 'PGText
       , userBio :: Col f ('PGNull PGText)
       }

In this example, each record (Package and User) correspond to tables in a PostgreSQL database. These records are parameterized over functors f, which will provide meaning depending on how the record is used. Notice all that we specify the types as they are in PostgreSQL. opaleye-tf primarily focuses on the flow of information from the database.

One type of meaning that we can give to tables is to map their fields to their respective columns. This is done by choosing Table as our choice of f:

packageTable :: Package (Table "package")
packageTable = Package { packageName = "name"
                       , packageAuthor = "author_id"
                       , packageMaintainerId = "maintainer_id" }

userTable :: User (Table "user")
userTable = User { userId = "id"
                 , userName = "name"
                 , userBio = "bio"
                 }

Now that we have full definitions of our tables, we can perform some SELECTs. First, let's list all known packages:

listAllPackages :: Connection -> IO [Package Interpret]
listAllPackages c = select (queryTable packageTable)

This computation now returns us a list of Packages where f has been set as Interpret. The Interpret functor is responsible for defining the mapping from PostgreSQL types (such as PGText) to Haskell types (such as Text).

Another choice of f occurs when we perform a left join. For example, here is the query to list all packages with their (optional) maintainers:

listAllPackagesAndMaintainersQuery :: Query (Package Expr, User (Compose Expr PGNull))
listAllPackagesAndMaintainersQuery =
  leftJoin (p m -> packageMaintainerId p ==. userId m)
           (queryTable package)
           (queryTable user))

This query communicates that we will have a collection of Expressions that correspond to columns in the Package table, and also a collection of Expressions that are columns in the User table. However, as the user table is only present as a left join, all of these columns might be NULL - indicated by the composition of PGNull with Expr.

opaleye-tf is smart enough to collapse data together where convenient. For example, if we look at just the final types of userId and userBio in the context of the left join:

> :t fmap (\(_, u) -> userId u) listAllPackagesAndMaintainersQuery
Query (Expr (PGNull PGInteger))

> :t fmap (\(_, u) -> userBio u) listAllPackagesAndMaintainersQuery
Query (Expr (PGNull PGText))

Which are as expected.

Finally, when executing queries that contain left joins, opaleye-tf is able to invert the possible NULLs over the whole record:

> :t conn -> select conn listAllPackagesAndMaintainersQuery
Connection -> IO [(Package Interpret, Maybe (User Interpret))]

Notice that User (Compose Expr PGNull) was mapped to Maybe (User Interpret).

type family Col f x :: b

The workhorse type family in opaleye-tf. This type family "collapses" noisy types down into things that are more meaningful. For example, when used with Interpret you just get ordinary Haskell values.

Equations

Col Interpret x = Apply InterpretPGType x

Col Insertion (PGDefault t) = Default (Expr t)

Col Insertion x = Expr x

Col (Compose Interpret PGNull) (PGNull t) = Maybe (Apply InterpretPGType t)

Col (Compose Interpret PGNull) x = Maybe (Col Interpret x)

Col (Compose Expr PGNull) (PGNull t) = Col (Compose Expr PGNull) t

Col (Compose Expr PGNull) t = Col Expr (PGNull t)

Col f x = f x

Mapping PostgreSQL types

data PGTypes

The universe of types known about by PostgreSQL.

Note that this data type is a little too large, in that it's possible to have types such as PGNull (PGNull PGInteger) which doesn't make much sense, so you'll need to be slightly careful!

Constructors

PGBigint

PGText

PGBoolean

PGNull PGTypes

PGReal

PGTimestampWithoutTimeZone

PGInteger

PGDefault PGTypes
Used to indicate that columns may also take an extra value - DEFAULT.

class (haskell ~ Apply InterpretPGType pg) => Lit pg haskell | haskell -> pg, pg -> haskell where

A class witnessing the relationship between PostgreSQL types and their Haskell representations.

Methods

lit :: haskell -> Expr pg

Given a Haskell value, build an Expr that represents this value as a PostgreSQL literal.

Defining tables

newtype Table tableName columnType

Table is used to specify the schema definition in PostgreSQL. The type itself uses a GHC "symbol" to specify the table name. This is also an instance of FromString, where the string is taken as the column name.

Example:

userTable :: User (Table "user")
userTable = User { userId = "id"
                 , userName = "name"
                 , userBio = "bio"
                 }

Constructors

Column String

Querying tables

queryTable :: forall table tableName. (KnownSymbol tableName, TableLike table tableName) => table (Table tableName) -> Query (table Expr)

queryTable moves from Table to Expr. Accessing the fields of your table record will now give you expressions to view data in the individual columns.

data Expr columnType

A single PostgreSQL expression of a given type.

select :: Selectable pg haskell => Connection -> Query pg -> IO [haskell]

select executes a PostgreSQL query as a SELECT statement, returning data mapped to Haskell values.

leftJoin :: ToNull right nullRight => (left -> right -> Expr PGBoolean) -> Query left -> Query right -> Query (left, nullRight)

A left join takes two queries and performs a relational LEFT JOIN between them. The left join has all rows of the left query, joined against zero or more rows in the right query. If the join matches no rows in the right table, then all columns will be null.

restrict :: QueryArr (Expr PGBoolean) ()

Apply a WHERE restriction to a table.

(==.) :: Expr a -> Expr a -> Expr PGBoolean infix 4

The PostgreSQL = operator.

Inserting data

insert :: Insertable table row => Connection -> table -> [row] -> IO Int64

Given a Table and a collection of rows for that table, INSERT this data into PostgreSQL. The rows are specified as PostgreSQL expressions.

data Insertion :: k -> *

Insertion witnesses that a table is being INSERTed. This adds special meaning to PGDefault'.

data Default a

Default is like Maybe, but only has meaning in INSERT statements.

Constructors

Default
Use the DEFAULT value for this column.

ProvideValue a
Override the default value by providing an explicit value.

Implementation details

data Compose f g a

Constructors

Compose (f (g a))

data InterpretPGType :: TyFun PGTypes * -> *

A defunctionalisation symbol used for mapping of database types to Haskell types.

data InterpretNull :: k -> *

data Interpret :: k -> *